backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pymysql==1.1.0
bcrypt==4.1.1
PyJWT==2.8.1
python-dotenv==1.0.0
